fix: vault npcs not appearing after server startup
